move time-range
