+++ src/time-range/index.ts
@@ -1,0 +1,2 @@
+export { TimeRange } from './time-range';
+export type { TimeSection } from './time-range.interface';
