+++ src/time-range/time-range.interface.ts
@@ -1,0 +1,5 @@
+export interface TimeSection {
+  start: number;
+  end: number;
+  duration: number;
+}
