+++ src/time-range/time-range.spec.ts
@@ -1,0 +1,128 @@
+import { MiscUtil } from '../misc-util';
+import { TimeRange } from './time-range';
+import { TimeSection } from './time-range.interface';
+
+describe('TimeRange Util', () => {
+  describe('time-range module', () => {
+    describe('check Constructor', () => {
+      it('start interval 1, start+10 = end, merge to one', async () => {
+        const loadSection: Array<TimeSection> = [];
+        for (let i = 0; i < 10; i++) {
+          const o: TimeSection = {
+            start: i,
+            end: i + 10,
+            duration: 10,
+          };
+          loadSection.push(o);
+        }
+        const timeRange = new TimeRange(loadSection);
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.value().length).toEqual(1);
+      });
+    });
+
+    describe('gathering time section', () => {
+      it('start interval 1, start+10 = end, merge to one', async () => {
+        const timeRange = new TimeRange();
+        for (let i = 0; i < 10; i++) {
+          const o: TimeSection = {
+            start: i,
+            end: i + 10,
+            duration: 10,
+          };
+          timeRange.add(o);
+        }
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.value().length).toEqual(1);
+      });
+
+      it('start interval 10, start+10 = end, merge to one', async () => {
+        const timeRange = new TimeRange();
+        for (let i = 0; i < 10; i++) {
+          const o: TimeSection = {
+            start: i * 10,
+            end: i * 10 + 10,
+            duration: 10,
+          };
+          timeRange.add(o);
+        }
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.value().length).toEqual(1);
+      });
+
+      it('start interval 10, start+10 = end, merge to two', async () => {
+        const timeRange = new TimeRange();
+        for (let i = 0; i < 4; i++) {
+          const o: TimeSection = {
+            start: i * 10,
+            end: i * 10 + 10,
+            duration: 10,
+          };
+          timeRange.add(o);
+        }
+        //split!
+        for (let i = 5; i < 11; i++) {
+          const o: TimeSection = {
+            start: i * 10,
+            end: i * 10 + 10,
+            duration: 10,
+          };
+          timeRange.add(o);
+        }
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.value().length).toEqual(2);
+      });
+
+      it('not ordering, merge to one', async () => {
+        const timeRange = new TimeRange();
+        const testArray = [1, 3, 0, 2, 7, 8, 5, 6, 9, 4];
+        for (const i of testArray) {
+          const o: TimeSection = {
+            start: i * 10,
+            end: i * 10 + 10,
+            duration: 10,
+          };
+          timeRange.add(o);
+        }
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.value().length).toEqual(1);
+      });
+
+      it('not ordering, uncertain interval, check totalDuration', async () => {
+        const timeRange = new TimeRange();
+        const testArray = [1, 3, 0, 2, 7, 8, 5, 6, 9, 4];
+        let totalDuration = 0;
+        for (const i of testArray) {
+          const r = MiscUtil.getRandomInt(1, 10);
+          const o: TimeSection = {
+            start: i * r,
+            end: i * r + r,
+            duration: r,
+          };
+          timeRange.add(o);
+          totalDuration += r;
+        }
+        expect(timeRange.value().length).toEqual(10);
+
+        timeRange.merge(true);
+
+        expect(timeRange.totalDuration()).toEqual(totalDuration);
+      });
+    });
+  });
+});
